+++ src/webdav/Resource.py
@@ -47,7 +47,7 @@ from ZPublisher.HTTPRangeSupport import HTTPRangeInterface
 from zope.interface import implements
 from zope.event import notify
 from zope.lifecycleevent import ObjectCopiedEvent
-from zope.container.contained import ObjectMovedEvent
+from zope.lifecycleevent import ObjectMovedEvent
 from zope.container.contained import notifyContainerModified
 
 from webdav.Lockable import LockableItem